mote TimeStepIter into time_step module
