+++ datatypes/src/util/mod.rs
@@ -4,4 +4,3 @@ pub mod ranges;
 mod result;
 pub mod test;
 pub use result::Result;
-pub mod time_step_iter;
